+++ CHANGELOG.md
@@ -7,4 +7,5 @@ Added source code, translated from JavaScript in [sql-formatter](https://github.
 Added customized tests.
 
 ## 2020-06-25
-Added constraints to not add `\n` before reserved keyword `AND` if the previous keyword is `BETWEEN`.
+1. Added constraints to not add `\n` before reserved keyword `AND` if the previous keyword is `BETWEEN`.
+2. Added blank line between sub-queries. The number of blank lines is customizable via `-linesBetweenQueries`.